// src/codex/generated/v2/DynamicToolSpec.ts
// GENERATED CODE! DO NOT MODIFY BY HAND!

// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DynamicToolFunctionSpec } from "./DynamicToolFunctionSpec.js";
import type { DynamicToolNamespaceSpec } from "./DynamicToolNamespaceSpec.js";

export type DynamicToolSpec =
  | ({ type: "function" } & DynamicToolFunctionSpec)
  | ({ type: "namespace" } & DynamicToolNamespaceSpec);